add error management on task ingest
